fix(sessions): address PR #211 review — safe submit, conflict regex, badge types, utils extraction
- Move payload construction inside try block so finally always runs setIsSubmitting(false)
- Fix isConflictError regex: add "conflito" (PT), replace generic "já existe" with "horário de sessão"
- Extract all session utility fns to session-utils.ts; update component and test imports
- Type badge tone maps with Record<Exclude<CatalogFormat, "">, BadgeTone> for exhaustiveness
- Add backend boundary test: rejects end_time 6 min short of runtime (outside 5-min tolerance)
- Fix misleading TZ comment in splitLocalDateTime test
